Time: 0 ms (100%), Space: 7.8 MB (30.14%) - LeetHub

+++ 0693-binary-number-with-alternating-bits/0693-binary-number-with-alternating-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool hasAlternatingBits(int n) {
-        unsigned int x = n ^ (n >> 1);
-        return (x & (x + 1)) == 0;
+        long x = n ^ (n >> 1);
+        return (x & (x + 1)) == 0;  // ✅ x+1 safe now, no overflow
     }
 };
